refactor(extension-common): content script のハンドラを移す
クレデンシャルの取得、Target Integrity の検証、広告リンクのクリック検知、
準備完了の通知を setupFrameHandlers() へ、Site Profile の取得を
setupTopFrameHandlers() へ移す。

フレーム CAS の位置算出はビュー要件で分岐しうるため、エントリ側に残した。
content-script-all-frames は 396 行から 105 行になった。
